Put some order in sbtc: the sbt compiler
The sbt compiler is just the wrapper around Scalac. I've decided to call
it sbtc.

This commit puts some order into the code by:

* Modularizing the functions in separate traits.
* Adding documentation to the core methods and sbt logic definitions.
* Adds comments to explain what's happening.
* Reorder the imports.

The way the sbtc code has been split is:

* SbtParser - For all the utils related to parsing.
* SbtHashGen - For the hashing utilities.
* SbtCodegen - For the codegen utilities.
* SbtDefinitions - For the domain-specific definitions.

There is no change in behaviour.

There is one change affecting the public API: `eval`, which is
deprecated in favour of `evalSetting`. `eval` is a misnomer since you
can evaluate both settings and definitions, and from the name it's not
clear that eval is a method to evaluate only settings. The new `eval`
forwards to `evalSetting` as expected.

There is one change affecting the internal API: rename of `valNames`
to `names` in `EvalDefinitions`. This change was done because the
previous name implies that the names are only for the vals, while sbt
expressions can also be `lazy val` and `def`s.