Rename `TrackedEvalResult`
`TrackedEvalResult` is the class that is used to represent an evaluation
result. The problem is that the prefix `Tracked` was added because
before it was not tracking the generated class files. However, I see no
reason why this should be encoded in the name of this class.

This commit renames it to something more sensible and predictable,
`EvalResult`. Some may think this is an unnecessary change, but I find
it very misleading that we clutter the namespace of such an important
logic in sbt: evaluation. The problem with these names is that they are
so augmented with superfluous details that makes it difficult to become
familiar with what they actually are.