Rename `target` to fix leaking test abstraction
It happens that the parameter `file` in `evaluateDefinitions` and
`evalDefinitions` does not only have a misleading name but it has an
incorrect type `Option[File]`. There is no way in sbt to evaluate
definitions that are not present in a sbt file, since you cannot create
a sbt definition from the sbt shell.

Why was this done, then?

Well, it was done because:

1. `EvalTest` didn't want to pass in a target directory; or
2. Some time ago, definitions could be evaluated in the sbt shell.

This commit takes care of:

1. Deprecating the existing method and forwarding to a new method that
   does not have the optional type.
2. Renaming `file` to `target`.
2. Updating the docs.