Evaluate sbt files at once instead of step by step
Previous to this commit, sbt evaluation logic worked as follows:

1. Compile and load definitions.
2. For every top-level expression, wrap it in an object, compile it and
   load it.

This encoding, however, is far than ideal and leads to ad-hoc logic and
possibly a runtime overhead. We could package and compile all together
while **keeping the same semantics**. A plus benefit of doing it this
way is that we will only load a class per sbt file, instead of different
classes per expressions + a class per definitions.

This commit addresses this issue, creating a `evalSbtFile` and
`evaluateSbtFileAtOnce` that replaces the default ones.