Move `appmacro` and `collection` to sbt/sbt
This commit moves everything behind the old `util-appmacro` and
`util-collection` over to sbt/sbt.

The reason for this change is because the logic of these modules is
really tight to the implementation of sbt and, I would even say, define
what sbt essentially is. They don't deserve to be called `util` nor to
be in an independent repository.

I have renamed the modules because I do not think they conveyed how
important these are for the sbt project. I don't have strong opinions
here, so I'm happy to pick other names if others consider appropriate.

`appmacro` has been renamed as `core-macros`. It is indeed the project
that defines the main logic to manipulate macros, and defines interfaces
that are then implemented by `main-settings`. Because of this strong
relationship between these two modules, I find that they have to stay
together and be able to be modified in one commit. I've realised this
while hacking on the macro changes for the usability of sbt.

`util-collection` has been renamed as `core-settings`. It defines all
the core structures (like `Init`, `AttributeKey`, `Setting`) that are
the core of sbt. These data structures are used by `appmacro` and the
same argument holds to move this module here. Given the importance of
these data structures, I feel that this should be one of the major
projects of sbt.