Call the soon-to-exist Global#close when done compiling
Scala 2.13.3 is likely to have a new method, `Global.close`.
This will let the compiler release resources, notably file
handles to JARs on the classpath.

This commit calls changes the compiler interfaces to call this
method, falling back to a no-op extension method for backwards
compatibility.

Furthermore, it changes the API to the global cache to let
the caller hand back the compiler instance. By default, this
compiler cache is a no-op (`GlobalCache.fresh`), and this is
where I've added the `close` call.

The other code path does not close the Global. AFAIK, caching
global instances (enabled with `sbt -Dsbt.resident.limit=<N>`)
a unfinished experiment and and not use in the wild. For instance,
I don't see how the current implementation ensures single threaded
access to the cached global instances, I would have expected a
check-out/check-in protocol to deal with that.

In any case, that code path is untouched under this patch.