No need for util-collection on Scala 2.10 anymore
We still publish parts of sbt for Scala 2.10 for compiler-bridge reasons
in Zinc.  However util-collection is now becoming an sbt-only module, so
it won't need to continue to cross-build to Scala 2.10!  Its new
dependency util-position which is used by lm also isn't a dependency of
the compiler-bridge, so also doesn't need to be cross-built to Scala
2.10.