Add more guarantees that a same module can't be downloaded concurrently
traverse called in ResolutionProcess.fetchAll relies on Applicative, so
doesn't guarantee that the module groups will be fetched one after the
other. The Applicative instance of scalaz.concurrent.Task doesn't
parallelize the tasks by default, so it works fine here. But that extra
security ensures that code can be fine with other monads.