Make slash syntax constructs eligible for .value enrichments
The implicit conversions in Def that add the .value, .evaluated,
.taskValue, .inputTaskValue and .parsed methods enrich Initialize, which
ScopeAndKey wasn't a subtype of.

Because of the lack of, what I believe is called, "kind polymorphism"
the fix requires an implementation for each of the 3 key type kinds.

Notably missing is the .previous enrichment, which is only on TaskKeys,
so I'm deferring the fix for .previous on slash syntax to a later date.

Fixes #3605