Bump Ivy to 2.3.0-sbt-1b57d3bbc08ecf671169fd548918da18c91f77be
Ref https://github.com/sbt/ivy/pull/26
Ref sbt/launcher#38

JDK 9 complains when java.lang.reflect.Field#setAccessible is called. So on
JDK 9 to get the default Authenticator on JDK you can just call
Authenticator.getDefault(). However ivy targets JDK 6..

So on JDK < 9 we do what we've always done, and on JDK 9+ we use java
reflection to call Authenticator.getDefault(), which is public and doesn't
require setAccessible(true).