Store setting values as key -> scope -> value
This helps a lot with performance mainly because during building the aggregation
index a huge number of `aggregate in X` values are looked up that for which the
whole delegate chain needs to be traversed until finally the default value in Global
is found.

Probably, there are even better ways to organize the storage but this
way will allow to optimize the main case, when there's only a single
value defined per key which already is a huge win.