Get rid of Scala collection overhead in hot scope calculation method
Nested flatMap doesn't optimize well, unfortunately.

There are probably even better ways of organizing delegation in a way
that it isn't necessary to generate the whole delegation chain if only
part of it is needed. In any case, the previous usage of `toStream` in
Settings0 only made things a lot worse because of additional allocation
and indirection.