Remove collectAnalyses step in sbt server initialize.
A client should trigger this command manually instead. Currently,
this automatic step triggers compilation in all root-aggregated
projects, which for large projects may take a long time to run.
It is better left to the client to decide what exact command to run on
startup, for example it might be desirable to scope the command by
custom project/config.